finished styling for weather
